+++ CHANGELOG.md
@@ -5,5 +5,6 @@
     * Fix: Updated Jigoshop logotypes.
     * Fix: Admin Order and Product list search.
     * Fix: Fatal error on admin order item variable template.
+    * Fix: Paypal response handling.
 * 2.0.0 - 2016.10.27:
     * Full rewrite of core.